Add changelogs on version update
Accidentally removed the toolbar in MainActivity. So far I haven't found the reason, therefore let's call it a feature and restructure the layout with a SlidingDrawer and BottomNavigation in a future release.

Also fixed the major fuck up of renaming "item" to "event" in resource files (Fuck you JetBrains refactoring!!!).