Add map fragment
This map shows the last occurred events in relation to the user current location. The user will be able to filter these events based on given conditions.

If an event is clicked, the events name (currently placeholder) and occurrence time are shown. If the user clicks the info window, a detail view of the event is shown.