Overhaul Google Maps integration
The lifecycle of GoogleMap is now respected and the current location is only updated as long the user views the map, which also fixes #11. Added my location button and removed current location marker, because GoogleMap now shows the well-known blue dot on your location.

The location requests in MapFragment are now set to high accuracy priority with an interval of 5 to 10 seconds. These values need to be investigated and tweaked to accomplish a fair trade off between location accuracy and battery consumption.

The permission system is simplified, because android.permission.INTERNET and android.permission.ACCESS_NETWORK_STATE are already in the GoogleMap manifest. Furthermore android.permission.ACCESS_COARSE_LOCATION is removed, because it's a subset of android.permission.ACCESS_FINE_LOCATION. Permission checks are also adapted to this change.

ButterKnife unbinds views in onDestroyView() in fragments. This reduces the memory usage a lot.

Major code cleanup and refactoring of values and layouts. Removed a lot of obsolete resources.