Change JSON deserializer from Gson to Moshi
Moshi will get Kotlin support, has better exception messages and is smaller (relies on Okio, which is already used). Therefore the model has changed. Read more about these changes in the upcoming release.

Also added FirebaseAuth for future use and set FirevbaseAnalytics screen names for current Activity/Fragment.

The BottomNavigation in MainActivity now uses an selector as item icon tint and item text color, to be able to show the user that an item is disabled.