Change event storage, location updates and configuration changes
Events are now stored in MainActivity and can be fetched by Fragments. This allows potential filtering in the future and avoids passing too much data.

Location updates are now only handled by MainActivity and Fragments can fetch the last location. This improves memory usage and avoids memory leaks and huge battery drain.

Several optimizations to the handling of configuration changes have also been made. The current selected item of the BottomNavigationView is now saved and the Fragment that was activated before will now be selected after configuration changes. There's still a lot of work to do, but i wanted to push this ASAP.