Add experimental Event features
Events are now separated between EventsFragment and MapFragment. MapFragment loads the events as usual from the server, but EventsFragment generates them locally and keeps them in it's RecyclerView's EventAdapter.

This not only separates the layers (MVP/MVVP incoming with Architecture Components), but also enables (direct) performance boosts by utilizing the DiffUtil, which is used to compare two sets of data. In this example the RecyclerView  uses a SwipeRefreshLayout to load new Events.

Sorting is currently not available and all event providing mechanisms (loading, generating) are moved to EventProvider.