Remove manual screen view tracking with Firebase Analytics
Analytics automatically tracks some information about screens in your application, such as the class name of the UIViewController or Activity that is currently in focus. When a screen transition occurs, Analytics logs a screen_view event that identifies the new screen. Events that occur on these screens are automatically tagged with the parameter firebase_screen_class (for example, menuViewController or MenuActivity) and a generated firebase_screen_id. If your app uses a distinct UIViewController or Activity for each screen, Analytics can automatically track every screen transition and generate a report of user engagement broken down by screen.